requirements.txt
lintrunner~=0.12.7
lintrunner-adapters~=0.12.4
torch~=2.6.0
timm~=1.0.15
nystrom_attention~=0.0.12
pandas~=2.2.3
wandb~=0.19.8
torcheval~=0.0.7
scikit-learn~=1.6.1
scikit-survival~=0.24.0
opencv-python~=4.11.0.86
opencv-contrib-python~=4.11.0.86
openslide-bin~=4.0.0.6
openslide-python~=1.4.1
scikit-image~=0.25.2
transformers~=4.49.0
albumentations~=2.0.5